refactor: move bottom sheet gesture logic from useEffect into a custom hook
